Prevent side window closure on pno form or pno reporting form dirty
