+++ frontend/src/features/Reporting/components/ReportingForm/index.tsx
@@ -1,82 +1,45 @@
 import { WindowContext } from '@api/constants'
-import { useGetControlUnitsQuery } from '@features/ControlUnit/controlUnitApi'
 import { CreateOrEditReportingSchema } from '@features/Reporting/components/ReportingForm/schemas'
-import {
-  getFormFields,
-  getReportingValue,
-  updateReportingActor
-} from '@features/Reporting/components/ReportingForm/utils'
-import { sortArrayByColumn } from '@features/Vessel/components/VesselList/tableSort'
+import { getFormFields, getReportingValue } from '@features/Reporting/components/ReportingForm/utils'
 import { useMainAppDispatch } from '@hooks/useMainAppDispatch'
-import { useMainAppSelector } from '@hooks/useMainAppSelector'
-import {
-  Accent,
-  Button,
-  FormikMultiRadio,
-  FormikSelect,
-  FormikTextarea,
-  FormikTextInput,
-  getOptionsFromLabelledEnum,
-  MultiRadio
-} from '@mtes-mct/monitor-ui'
-import { Form, Formik } from 'formik'
-import { useCallback, useMemo } from 'react'
-import styled from 'styled-components'
+import { Formik } from 'formik'
+import { useCallback, useEffect } from 'react'
 
+import { Form } from './Form'
 import { getOnlyVesselIdentityProperties } from '../../../../domain/entities/vessel/vessel'
-import {
-  ReportingOriginActor,
-  ReportingOriginActorLabel,
-  ReportingType,
-  ReportingTypeCharacteristics
-} from '../../types'
 import { addReporting } from '../../useCases/addReporting'
 import { updateReporting } from '../../useCases/updateReporting'
-import { mapControlUnitsToUniqueSortedIdsAsOptions } from '../VesselReportingList/utils'
 
 import type { VesselIdentity } from '../../../../domain/entities/vessel/types'
-import type { EditedReporting, InfractionSuspicion, Reporting } from '../../types'
-import type { Option } from '@mtes-mct/monitor-ui'
+import type { EditedReporting, Reporting } from '../../types'
 
 type ReportingFormProps = {
   className?: string | undefined
-  closeForm: () => void
   editedReporting: Reporting.EditableReporting | undefined
   hasWhiteBackground?: boolean
+  onClose: () => void
+  onIsDirty?: ((isDirty: boolean) => void) | undefined
   vesselIdentity: VesselIdentity
   windowContext: WindowContext
 }
 export function ReportingForm({
   className,
-  closeForm,
   editedReporting,
   hasWhiteBackground = false,
+  onClose,
+  onIsDirty,
   vesselIdentity,
   windowContext
 }: ReportingFormProps) {
   const dispatch = useMainAppDispatch()
-  // TODO Replace that with a `useInfractionsAsOptions()` hook with RTK query.
-  const infractions = useMainAppSelector(state => state.infraction.infractions)
-  const controlUnitsQuery = useGetControlUnitsQuery(undefined)
 
-  const infractionsAsOptions = useMemo(
-    () =>
-      infractions
-        .map(infraction => ({
-          label: `${infraction.natinfCode} - ${infraction.infraction}`,
-          value: infraction.natinfCode
-        }))
-        .sort((a, b) => sortArrayByColumn(a, b, 'label', 'asc')),
-    [infractions]
-  )
-
-  const controlUnitsAsOptions = useMemo((): Option<number>[] => {
-    if (!controlUnitsQuery.data) {
-      return []
+  const handleClose = useCallback(() => {
+    if (onIsDirty) {
+      onIsDirty(false)
     }
 
-    return mapControlUnitsToUniqueSortedIdsAsOptions(controlUnitsQuery.data)
-  }, [controlUnitsQuery.data])
+    onClose()
+  }, [onClose, onIsDirty])
 
   const createOrEditReporting = useCallback(
     async (reportingValue: EditedReporting) => {
@@ -93,7 +56,7 @@ export function ReportingForm({
           )
         )
 
-        closeForm()
+        handleClose()
 
         return
       }
@@ -116,9 +79,18 @@ export function ReportingForm({
 
       dispatch(addReporting(nextReportingWithMissingProperties))
 
-      closeForm()
+      handleClose()
     },
-    [dispatch, closeForm, editedReporting, vesselIdentity, windowContext]
+    [dispatch, editedReporting, handleClose, vesselIdentity, windowContext]
+  )
+
+  useEffect(
+    () => () => {
+      if (onIsDirty) {
+        onIsDirty(false)
+      }
+    },
+    [onIsDirty]
   )
 
   return (
@@ -127,146 +99,13 @@ export function ReportingForm({
       onSubmit={createOrEditReporting}
       validationSchema={CreateOrEditReportingSchema}
     >
-      {({ setFieldValue, values }) => {
-        const updateActor = updateReportingActor(setFieldValue)
-        const infractionTitle = infractions?.find(
-          infraction => infraction.natinfCode === (values as Partial<InfractionSuspicion>).natinfCode
-        )?.infraction
-
-        return (
-          <StyledForm
-            $hasWhiteBackground={hasWhiteBackground}
-            $isInfractionSuspicion={values.type === ReportingType.INFRACTION_SUSPICION}
-            className={className}
-          >
-            <FormikMultiRadio
-              isInline
-              isLight={!hasWhiteBackground}
-              label="Type de signalement"
-              name="type"
-              options={[
-                {
-                  label: ReportingTypeCharacteristics.INFRACTION_SUSPICION.inputName,
-                  value: ReportingTypeCharacteristics.INFRACTION_SUSPICION.code
-                },
-                {
-                  label: ReportingTypeCharacteristics.OBSERVATION.inputName,
-                  value: ReportingTypeCharacteristics.OBSERVATION.code
-                }
-              ]}
-            />
-            <MultiRadio
-              isInline
-              isLight={!hasWhiteBackground}
-              label="Origine"
-              name="reportingActor"
-              onChange={updateActor}
-              options={getOptionsFromLabelledEnum(ReportingOriginActorLabel)}
-              value={values.reportingActor}
-            />
-            {values.reportingActor === ReportingOriginActor.UNIT && (
-              <FormikSelect
-                isLight={!hasWhiteBackground}
-                label="Choisir l'unité"
-                name="controlUnitId"
-                options={controlUnitsAsOptions}
-                searchable
-              />
-            )}
-            {(values.reportingActor === ReportingOriginActor.UNIT ||
-              values.reportingActor === ReportingOriginActor.DML ||
-              values.reportingActor === ReportingOriginActor.DIRM ||
-              values.reportingActor === ReportingOriginActor.OTHER) && (
-              <FormikTextInput
-                isLight={!hasWhiteBackground}
-                label="Nom et contact (numéro, mail…) de l’émetteur"
-                name="authorContact"
-                placeholder="Ex: Yannick Attal (06 24 25 01 91)"
-              />
-            )}
-            <FormikTextInput
-              isLight={!hasWhiteBackground}
-              label="Titre"
-              name="title"
-              placeholder={
-                values.type === ReportingTypeCharacteristics.OBSERVATION.code
-                  ? 'Ex: Dérogation temporaire licence'
-                  : 'Ex: Infraction maille cul de chalut'
-              }
-            />
-            <FormikTextarea
-              isLight={!hasWhiteBackground}
-              label="Description"
-              name="description"
-              placeholder={
-                values.type === ReportingTypeCharacteristics.OBSERVATION.code
-                  ? "Ex: Licence en cours de renouvellement, dérogation accordée par la DML jusqu'au 01/08/2022."
-                  : 'Ex: Infraction constatée sur la taille de la maille en cul de chalut'
-              }
-            />
-            {values.type === ReportingTypeCharacteristics.INFRACTION_SUSPICION.code && (
-              <FormikSelect
-                isLight={!hasWhiteBackground}
-                label="Natinf"
-                name="natinfCode"
-                options={infractionsAsOptions}
-                placement={windowContext === WindowContext.MainWindow ? 'topStart' : undefined}
-                searchable
-                // @ts-ignore
-                title={infractionTitle}
-              />
-            )}
-            <FormikTextInput
-              isLight={!hasWhiteBackground}
-              label="Saisi par"
-              name="authorTrigram"
-              placeholder="Ex: LTH"
-            />
-            <ValidateButton accent={Accent.PRIMARY} type="submit">
-              Valider
-            </ValidateButton>
-            <CancelButton accent={Accent.SECONDARY} onClick={closeForm}>
-              Annuler
-            </CancelButton>
-          </StyledForm>
-        )
-      }}
+      <Form
+        className={className}
+        hasWhiteBackground={hasWhiteBackground}
+        onClose={handleClose}
+        onIsDirty={onIsDirty}
+        windowContext={windowContext}
+      />
     </Formik>
   )
 }
-
-const ValidateButton = styled(Button)`
-  margin: 24px 10px 0px 0px;
-`
-
-const CancelButton = styled(Button)`
-  margin: 24px 0px 0px 0px;
-`
-
-const StyledForm = styled(Form)<{
-  $hasWhiteBackground: boolean
-  $isInfractionSuspicion: boolean
-}>`
-  background: ${p => {
-    if (p.$hasWhiteBackground) {
-      return 'unset'
-    }
-
-    return p.$isInfractionSuspicion ? p.theme.color.maximumRed15 : p.theme.color.gainsboro
-  }};
-  padding-right: 16px;
-  padding-left: 16px;
-
-  * {
-    box-sizing: border-box !important;
-  }
-
-  > .Element-Field,
-  > .Element-Fieldset:not(:first-child) {
-    margin-top: 16px;
-  }
-
-  > .Field-MultiRadio > legend {
-    margin-bottom: 8px;
-  }
-`
